Move action selectors to actions directory
